Change name of global cache dir to .winappglobal instead of .winapp (#178)
We have a test case that's intermittently failing, it turns out this is
because some of the tests set the global .winapp dir and local .winapp
dir to be the same. I added logic recently that assumed they could never
be the same, which was a bad assumption.

To avoid confusion over these dirs, let's rename the global cache dir to
.winappglobal.

---------

Co-authored-by: Alexandre Zollinger Chohfi <alzollin@microsoft.com>